+++ NEWS.md
@@ -1,5 +1,8 @@
 # insurancerating 0.7.1
 
+* `rotate_angle` in `autoplot.univariate()` is added to rotate x-labels
+* `univariate()` now accepts external vectors for `x`; `vec_ext()` must be used
+
 # insurancerating 0.7.0
 
 * `smooth_coef()` now gives correct results for intervals with scientific notation